Added login_required to edit/delete posts, search functionality. Added register page

+++ base/views.py
@@ -1,11 +1,13 @@
 from django.db.models import expressions
 from django.contrib import messages
+from django.http import HttpResponse
 from django.shortcuts import render, redirect
-# from django.contrib.auth.decorators import login_required
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth import authenticate, login, logout
-
-from .models import User, Room
-from .forms import RoomForm
+from django.contrib.auth.forms import UserCreationForm
+from django.db.models import Q
+from .models import User, Room, Topic
+from .forms import RoomForm, UserRegisterForm
 
 # Create your views here.
 
@@ -13,8 +15,18 @@ from .forms import RoomForm
 
 
 def home(request):
-    rooms = Room.objects.all()
-    context = {'rooms': rooms}
+    q = request.GET.get('q') if request.GET.get('q') != None else ''
+
+    rooms = Room.objects.filter(
+        Q(topic__name__icontains=q) |
+        Q(name__icontains=q) |
+        Q(description__icontains=q)
+        )
+
+    topics = Topic.objects.all()
+    room_count = rooms.count()
+
+    context = {'rooms': rooms, 'topics': topics, 'room_count': room_count}
     return render(request, 'base/home.html', context)
 
 def room(request, pk):
@@ -53,7 +65,19 @@ def logoutPage(request):
     logout(request)
     return redirect('home')
 
+def registerPage(request):
+    form = UserRegisterForm()
 
+    if request.method == 'POST':
+        form = UserRegisterForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.save()
+
+    return render(request, 'base/login.html', {'form': form})
+
+
+@login_required(login_url='login')
 def createRoom(request):
 
     form = RoomForm()
@@ -67,9 +91,13 @@ def createRoom(request):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here.')
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -81,8 +109,13 @@ def updateRoom(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here.')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
